refactor(simu): use C++ native threads
And remove direct dependency on POSIX threads.

# Conflicts:
#	radio/src/targets/simu/CMakeLists.txt
