+++ radio/src/os/timer_native_impl.h
@@ -1,4 +1,4 @@
-#include "timer_pthread.h"
+#include "timer_native.h"
 
 #include <condition_variable>
 #include <deque>
